fix: unify nodeclass status and termination controllers to prevent ra… (#7597)
