+++ pkg/controllers/nodeclass/events.go
@@ -12,7 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package termination
+package nodeclass
 
 import (
 	"fmt"
