+++ pkg/controllers/nodeclass/instanceprofile.go
@@ -12,7 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package status
+package nodeclass
 
 import (
 	"context"
@@ -40,5 +40,12 @@ func (ip *InstanceProfile) Reconcile(ctx context.Context, nodeClass *v1.EC2NodeC
 		nodeClass.Status.InstanceProfile = lo.FromPtr(nodeClass.Spec.InstanceProfile)
 	}
 	nodeClass.StatusConditions().SetTrue(v1.ConditionTypeInstanceProfileReady)
+	return reconcile.Result{}, nil
+}
+
+func (ip *InstanceProfile) Finalize(ctx context.Context, nodeClass *v1.EC2NodeClass) (reconcile.Result, error) {
+	if err := ip.instanceProfileProvider.Delete(ctx, nodeClass); err != nil {
+		return reconcile.Result{}, fmt.Errorf("deleting instance profile, %w", err)
+	}
 	return reconcile.Result{}, nil
 }
